Cast fread to int32_t

+++ src/lgca.c
@@ -566,7 +566,7 @@ void lecture_mvt() {
     exit(-4);
   }
 
-  if (fread(CelMvt, sizeof(MvtField), CHLD, fp) != CHLD) {
+  if ((int32_t)fread(CelMvt, sizeof(MvtField), CHLD, fp) != CHLD) {
     ErrPrintf("Read error\n");
     exit(-1);
   }
